@@ -3,7 +3,7 @@ name = "arithmetic-parser"
 version = "0.3.0"
 authors = ["Alex Ostrovski <ostrovski.alex@gmail.com>"]
 edition = "2021"
-rust-version = "1.60"
+rust-version = "1.65"
 readme = "README.md"
 license = "MIT OR Apache-2.0"
 description = "Parser for arithmetic expressions with flexible literals and type annotations."
@@ -17,7 +17,7 @@ all-features = true
 [dependencies]
 # Public dependencies (present in the public API).
 anyhow = { version = "1.0.27", default-features = false }
-bitflags = "1.2.1"
+bitflags = "2.1.0"
 nom = { version = "7", default-features = false, features = ["alloc"] }
 nom_locate = { version = "4", default-features = false, features = ["alloc"] }
 num-traits = { version = "0.2.14", default-features = false }
