@@ -24,9 +24,7 @@ impl NativeFn<f32> for PointLen {
         context: &mut CallContext<'_, 'a, f32>,
     ) -> EvalResult<'a, f32> {
         context.check_args_count(&args, 1)?;
-        let point = if let Value::Object(obj) = args.pop().unwrap().extra {
-            obj
-        } else {
+        let Value::Object(point) = args.pop().unwrap().extra else {
             let err = ErrorKind::native("Function argument must be an object");
             return Err(context.call_site_error(err));
         };
@@ -173,16 +171,14 @@ fn defining_prototype_in_script() {
     evaluate(&mut env, program);
 
     let proto_string = env["Manhattan"].to_string();
-    assert!(proto_string.contains("prototype"), "{}", proto_string);
+    assert!(proto_string.contains("prototype"), "{proto_string}");
     assert!(
         proto_string.contains("dist: (interpreted fn @ *:4:32)"),
-        "{}",
-        proto_string
+        "{proto_string}"
     );
     assert!(
         proto_string.contains("len: (interpreted fn @ *:3:19)"),
-        "{}",
-        proto_string
+        "{proto_string}"
     );
 }
 
