@@ -15,9 +15,8 @@ use crate::{evaluate, expect_compilation_error, try_evaluate};
 fn object_basics() {
     let program = "#{ x: 1, y: (2, 3) }";
     let return_value = evaluate(&mut Environment::new(), program);
-    let fields = match return_value {
-        Value::Object(fields) => fields,
-        _ => panic!("Unexpected return value: {return_value:?}"),
+    let Value::Object(fields) = return_value else {
+        panic!("Unexpected return value: {return_value:?}");
     };
 
     assert_eq!(fields.len(), 2);
@@ -45,9 +44,8 @@ fn capturing_vars_in_object() {
         #{ x, y: y + 1 }
     "#;
     let return_value = evaluate(&mut Environment::new(), program);
-    let fields = match return_value {
-        Value::Object(fields) => fields,
-        _ => panic!("Unexpected return value: {return_value:?}"),
+    let  Value::Object(fields) = return_value else {
+        panic!("Unexpected return value: {return_value:?}");
     };
 
     assert_eq!(fields.len(), 2);
@@ -64,9 +62,8 @@ fn object_expr_does_not_capture_surroundings() {
         obj
     "#;
     let return_value = evaluate(&mut Environment::new(), program);
-    let fields = match return_value {
-        Value::Object(fields) => fields,
-        _ => panic!("Unexpected return value: {return_value:?}"),
+    let Value::Object(fields) = return_value else {
+        panic!("Unexpected return value: {return_value:?}");
     };
 
     assert_eq!(fields.len(), 2);
@@ -84,9 +81,8 @@ fn object_expr_does_not_capture_inner_scopes() {
         }
     "#;
     let return_value = evaluate(&mut Environment::new(), program);
-    let fields = match return_value {
-        Value::Object(fields) => fields,
-        _ => panic!("Unexpected return value: {return_value:?}"),
+    let Value::Object(fields) = return_value else {
+        panic!("Unexpected return value: {return_value:?}");
     };
 
     assert_eq!(fields.len(), 3);
@@ -108,17 +104,15 @@ fn object_in_object() {
         }
     "#;
     let return_value = evaluate(&mut Environment::new(), program);
-    let fields = match return_value {
-        Value::Object(fields) => fields,
-        _ => panic!("Unexpected return value: {return_value:?}"),
+    let Value::Object(fields) = return_value else {
+        panic!("Unexpected return value: {return_value:?}");
     };
 
     assert_eq!(fields.len(), 2);
     assert_eq!(fields["pt_x"], Value::Prim(1.0));
 
-    let inner_fields = match &fields["pt"] {
-        Value::Object(fields) => fields,
-        other => panic!("Unexpected inner object: {other:?}"),
+    let Value::Object(inner_fields) = &fields["pt"] else {
+        panic!("Unexpected inner object: {fields:?}");
     };
     assert_eq!(inner_fields.len(), 2);
     assert_eq!(inner_fields["x"], Value::Prim(1.0));
@@ -146,9 +140,8 @@ fn accessing_fields_within_object() {
         #{ pt, len_sq: pt.x * pt.x + pt.y * pt.y }
     "#;
     let return_value = evaluate(&mut Environment::new(), program);
-    let fields = match return_value {
-        Value::Object(fields) => fields,
-        _ => panic!("Unexpected return value: {return_value:?}"),
+    let Value::Object(fields) = return_value else {
+        panic!("Unexpected return value: {return_value:?}");
     };
 
     let expected_fields = [("x", Value::Prim(3.0)), ("y", Value::Prim(4.0))];
