@@ -32,8 +32,7 @@ fn is_valid_variable_name_works() {
     for &invalid_name in &["", "1abc", "\u{43d}\u{435}\u{442}", "xy+", "a-b"] {
         assert!(
             !is_valid_variable_name(invalid_name),
-            "failed at invalid name: {}",
-            invalid_name
+            "failed at invalid name: {invalid_name}"
         );
     }
 }
@@ -94,11 +93,10 @@ fn multiline_comments() {
 
     let input = InputSpan::new("/* Foo");
     let err = ws::<Complete>(input).unwrap_err();
-    let err = match &err {
-        NomErr::Failure(err) => err.kind(),
-        _ => panic!("Unexpected error: {err:?}"),
+    let NomErr::Failure(err) = &err else {
+        panic!("Unexpected error: {err:?}");
     };
-    assert_matches!(err, ErrorKind::UnfinishedComment);
+    assert_matches!(err.kind(), ErrorKind::UnfinishedComment);
 }
 
 #[test]
